backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.4.2
httpx==0.25.0
python-dotenv==1.0.0
openai==1.2.3
spotipy==2.23.0
wikipedia-api==0.6.0
supabase==1.0.3
python-jose==3.3.0
celery==5.3.4
redis==5.0.1
pytest==7.4.2 